backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Pydantic for data validation
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP client for Chess.com API
httpx==0.26.0

# Redis client (async)
redis==5.0.1

# Database (PostgreSQL)
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25

# Chess engine
python-chess==1.999
stockfish==3.28.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.9

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0  # Already listed above, used for testing

# Development
black==24.1.1
ruff==0.2.1
mypy==1.8.0

# ASGI server
gunicorn==21.2.0
